# [IMP] base: faster search on filters by re-ordering unique constraint