--- a/odoo/addons/base/populate/__init__.py
+++ b/odoo/addons/base/populate/__init__.py
@@ -1,3 +1,4 @@
+from . import ir_filters
 from . import res_partner
 from . import res_company
 from . import res_currency